게시물 삭제시 권한체크

+++ src/main/java/com/example/lfcFan/controller/usr/ArticleController.java
@@ -125,6 +125,14 @@ public class ArticleController {
 			return "common/redirect";
 		}
 
+		Article article = articleService.getArticleById(id);
+
+		if ( article.getMemberId() != loginedMemberId ) {
+			model.addAttribute("msg", "권한이 없습니다.");
+			model.addAttribute("historyBack", true);
+			return "common/redirect";
+		}
+		
 		articleService.deleteArticleById(id);
 
 		model.addAttribute("msg", String.format("%d번 글을 삭제하였습니다.", id));
